Move terrain attributes to new package
